move functions from common_txn to txn_builder
